+++ CHANGELOG.md
@@ -4,6 +4,16 @@ Prepend new items to make git merging easier.
 - refactor(autocompleter cell): use a stylesheets API
 - refactor(input control): use Icons component for invalid state
 - chore(useMounted): deprecate useMounted hook since it does not work unless the useEffect is "last" in the render cycle
+
+## 0.100.3
+- fix(multi-select): don't set invalid when blurring to click an option from the dropdown
+
+## 0.100.1
+
+- refactor(multi-aucompleter): revert URL generation from browser API to static string
+
+## 0.100.0
+
 - fix(multi-select cell): add validation message as a tooltip on the invalid icon
 - fix(multi-autocompleter): do not call onChange when loading selected values on first-time render
 - fix(multi-choice control): always load provided selected values in case search pagination does not return a result set with the selected values
@@ -21,6 +31,9 @@ Prepend new items to make git merging easier.
 - refactor(multi-select): remove id attribute on container (since it is unused)
 - refactor(multi-select): move container styling inside the form control for easier refactoring to a multi-select control
 - refactor(multi-select): move dropdown inside the form control for easier refactoring to a multi-select control
+
+## 0.99.2
+
 - fix(choice control): only fetch active choices to avoid "duplicates" bug
 
 ## 0.99.1